Assignments/Quarter-5/LangGraph/pyproject.toml
[tool.poetry]
name = "langagent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
langgraph = "^0.2.59"
langchain = "^0.3.12"
python-dotenv = "^1.0.1"
langchain-huggingface = "^0.1.2"
langchain-community = "^0.3.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
